+++ examples/default-provider/app/(default)/custom-player/page.tsx
@@ -1,7 +1,7 @@
 import { Metadata } from 'next';
 import Video from 'next-video';
 import ReactPlayer from './player'
-import countryClouds from '/videos/country-clouds.mp4';
+import countryClouds from '../../../videos/country-clouds.mp4';
 
 export const metadata: Metadata = {
   title: 'next-video - Custom Player',