+++ examples/default-provider/next.config.mjs
@@ -1,5 +1,6 @@
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
+// import { withNextVideo } from 'next-video/process';
 import { withNextVideo } from './next-video.mjs';
 
 const fileDir = path.dirname(fileURLToPath(import.meta.url));
@@ -12,6 +13,14 @@ const nextConfig = (phase, { defaultConfig }) => {
     // https://github.com/vercel/next.js/issues/64472#issuecomment-2077483493
     // https://nextjs.org/docs/pages/api-reference/config/next-config-js/output#caveats
     outputFileTracingRoot: path.join(fileDir, '../../'),
+    images: {
+      remotePatterns: [
+        {
+          protocol: 'https',
+          hostname: 'image.mux.com',
+        },
+      ],
+    },
   };
 };